requirements.txt
anthropic>=0.40.0
pdfplumber>=0.11.0
chromadb>=0.5.0
jinja2>=3.1.0
click>=8.1.0
rich>=13.0.0
sentence-transformers>=3.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
Pillow>=10.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-multipart>=0.0.9
stripe>=7.0.0
# OCR for scanned PDF drawings (image-based pages)
pdf2image>=1.16.0
pytesseract>=0.3.10
# PostgreSQL + pgvector (optional — activate by setting DATABASE_URL or DB_HOST)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pgvector>=0.3.0
alembic>=1.13.0
